+++ tinper-bee/bee-timepicker/src/Timepicker.js
@@ -17,22 +17,34 @@ class Timepicker extends Component {
             value:props.value
         }
     }
+    
+    initValue=()=>{
 
-    componentDidMount(){
+    }
+
+    componentWillMount(){
         //判断初始值是否合法，不合法则格式化，并触发onChange回调
         let value = this.props.value;
         if(value){
             if(typeof value == 'string'){
                 value = moment(`${moment().format('YYYY-MM-DD')} ${value}`);
+                this.setState({
+                    value
+                })
+                this.props.onChange&&this.props.onChange(value)
             }else if(value.format){
                 if(!value.isValid()){
-                    value = moment(`${moment().format('YYYY-MM-DD')} ${value._i}`)
+                    value = moment(`${moment().format('YYYY-MM-DD')} ${value._i}`);
+                    this.setState({
+                        value
+                    })
+                    this.props.onChange&&this.props.onChange(value)
                 }
             }
         }else{
             value = null
         }
-        this.props.onChange&&this.props.onChange(value)
+        
     }
 
     componentWillReceiveProps(nextProps){