+++ tinper-bee/bee-calendar/CHANGELOG.md
@@ -1,3 +1,13 @@
+<a name="2.1.2"></a>
+## [2.1.2](https://github.com/tinper-bee/bee-calendar/compare/v2.1.1...v2.1.2) (2019-11-25)
+
+
+
+<a name="2.1.1"></a>
+## [2.1.1](https://github.com/tinper-bee/bee-calendar/compare/v2.1.0...v2.1.1) (2019-11-19)
+
+
+
 <a name="2.1.0"></a>
 # [2.1.0](https://github.com/tinper-bee/bee-calendar/compare/v2.0.4...v2.1.0) (2019-09-10)