+++ tinper-bee/bee-input-number/CHANGELOG.md
@@ -1,3 +1,13 @@
+<a name="2.2.3"></a>
+## [2.2.3](https://github.com/tinper-bee/bee-input-number/compare/v2.2.2...v2.2.3) (2019-11-21)
+
+
+
+<a name="2.2.2"></a>
+## [2.2.2](https://github.com/tinper-bee/bee-input-number/compare/v2.2.1...v2.2.2) (2019-11-18)
+
+
+
 <a name="2.2.1"></a>
 ## [2.2.1](https://github.com/tinper-bee/bee-input-number/compare/v2.2.0...v2.2.1) (2019-11-14)