+++ tinper-bee/bee-pagination/CHANGELOG.md
@@ -1,3 +1,13 @@
+<a name="2.0.14"></a>
+## [2.0.14](https://github.com/tinper-bee/bee-pagination/compare/v2.0.13...v2.0.14) (2019-11-21)
+
+
+
+<a name="2.0.13"></a>
+## [2.0.13](https://github.com/tinper-bee/bee-pagination/compare/v2.0.12...v2.0.13) (2019-11-14)
+
+
+
 <a name="2.0.12"></a>
 ## [2.0.12](https://github.com/tinper-bee/bee-pagination/compare/v2.0.11...v2.0.12) (2019-10-18)